upd: optimize school scraping

+++ src/Ellipse.Common/Models/Matrix/OpenRoute/OpenRouteMatrixResponse.cs
@@ -5,8 +5,8 @@ namespace Ellipse.Common.Models.Matrix.OpenRoute;
 public class OpenRouteMatrixResponse
 {
     [JsonPropertyName("durations")]
-    public double[][]? Durations { get; set; }
+    public float[][]? Durations { get; set; }
 
     [JsonPropertyName("distances")]
-    public double[][]? Distances { get; set; }
+    public float[][]? Distances { get; set; }
 }
